# internal/domain/customer.go
package domain

import "time"

type Customer struct {
	ID        string    `json:"id" bson:"id"`
	FirstName string    `json:"firstName" bson:"firstName"`
	LastName  string    `json:"lastName" bson:"lastName"`
	Email     string    `json:"email" bson:"email"`
	CreatedAt time.Time `json:"createdAt" bson:"createdAt"`
}